backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
firebase-admin==6.2.0
gunicorn==21.2.0
python-telegram-bot==20.7
aiohttp==3.9.1
asyncio==3.4.3 